feat(train/training_internlm.py): move common init funcs to internlm/train
